% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_op_periods}
\alias{add_op_periods}
\title{Add Operational Periods}
\usage{
add_op_periods(
  responses,
  startDate,
  endDate,
  period_length = 12,
  period_name = "Operational Period"
)
}
\arguments{
\item{responses}{Source data set with the DispatchedTime POSIXct field}

\item{startDate}{Date time of interval starts}

\item{endDate}{Date time of interval ends}

\item{period_length}{Length of each operational period in hours}

\item{period_name}{Name to insert in column}
}
\value{
responses + op_period field
}
\description{
Uses DispatchedTime field in responses dataset and matches
against operational period intervals. Adds a column named
"op_period" with the value of period_name, an indice, and the
interval in YYYY-MM-DD HH:MM:SS format.
}
\examples{
\dontrun{
startDate <- mdy_hms("02/14/2021 18:00:00", tz="US/Central")
endDate <- mdy_hms("02/19/2021 18:00:00", tz="US/Central")
add_op_periods(responses, startDate, endDate)
}
# adds op_period column with value of (e.g.) "Operational Period 1 (2021-02-14 18:00:00-2021-02-15 06:00:00)"

}
