% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{blood}
\alias{blood}
\title{Blood}
\usage{
blood(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{tibble of incidents}

\item{year}{Year of data set as "YYYY" string}

\item{month}{Month of data set as "MM" string, or FALSE to look for full year.}
}
\value{
tibble with Blood field
}
\description{
Transforms incidents table to include a field "Blood" indicating if the patient received blood products or not.
}
