% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{cardiac_arrest}
\alias{cardiac_arrest}
\title{Adds OHCA field to Incidents Table
Transforms incidents table to include a
field "OHCA" indicating if the patient
was a cardiac arrest or not.}
\usage{
cardiac_arrest(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{Incidents table}

\item{month}{}
}
\value{
Incidents + OHCA field
}
\description{
Adds OHCA field to Incidents Table
Transforms incidents table to include a
field "OHCA" indicating if the patient
was a cardiac arrest or not.
}
