% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{convert_estimated_age}
\alias{convert_estimated_age}
\title{Convert estimated age to decimal age in years}
\usage{
convert_estimated_age(patient_age)
}
\arguments{
\item{patient_age}{Age in string}
}
\value{
Double of patient age in years
}
\description{
Convert estimated age to decimal age in years
}
