% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma.R
\name{filter_unstable_vitals}
\alias{filter_unstable_vitals}
\title{Filter Unstable Vitals}
\usage{
filter_unstable_vitals(.data, patients, start, end)
}
\arguments{
\item{.data}{Incidents tibble we are augmenting.}

\item{patients}{Patients tibble}

\item{start}{Start Date}

\item{end}{End Date}
}
\value{
.data tibble with additional field for unstable vitals.
}
\description{
Augments an incidents tibble (with PatientCareRecordId primary key) with unstable vitals field.
}
