% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{filtered_incidents}
\alias{filtered_incidents}
\title{Imports a Filtered Incident Set
Imports incidents table filtered against the filter_set
passed to it. Filter Set has to have PatientCareRecordId index.}
\usage{
filtered_incidents(year, month, filter_set)
}
\arguments{
\item{filter_set}{Tibble of ESO data}
}
\description{
Imports a Filtered Incident Set
Imports incidents table filtered against the filter_set
passed to it. Filter Set has to have PatientCareRecordId index.
}
