% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freq_table}
\alias{freq_table}
\title{freq_table}
\usage{
freq_table(.data, group_var, summary = TRUE)
}
\arguments{
\item{.data}{tibble to use}

\item{group_var}{vector of columns to group by}

\item{summary}{logi whether to summarize}
}
\value{
summarized (or not) list
}
\description{
freq_table reports frequencies, percentages, and totals for a tibble
based on grouping columns.
}
