% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gems_plot.R
\name{gems_plot}
\alias{gems_plot}
\title{Generates a "Gems" style plot}
\usage{
gems_plot(data, main_plot, title = "")
}
\arguments{
\item{data}{Data coming in}

\item{main_plot}{Main plot to add gems to}

\item{title}{Title of plot}
}
\value{
plot format
}
\description{
Creates plot with data points on left side and a main plot on right.
Data points coming in are parsed as a single row with measure names
as the variable/column names, and measure values as the data.
}
