% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eso_files.R
\name{generate_filename}
\alias{generate_filename}
\title{Generates a file name to be used to get a path}
\usage{
generate_filename(year, month = FALSE, table_name)
}
\arguments{
\item{year}{YYYY formatted string}

\item{month}{MM formatted string or FALSE to omit}

\item{table_name}{standardized name of table}
}
\value{
filename
}
\description{
Generates a file name to be used to get a path
}
