% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{get_age_range}
\alias{get_age_range}
\title{Takes age in years and returns the age range}
\usage{
get_age_range(patient_age)
}
\arguments{
\item{patient_age}{Age in string}
}
\value{
age range definition as string
}
\description{
Ages as follows:
Newborn: < 24 hours
Neonate: 1 day - 1 month
Infant: 1 month - 1 year
Toddler: 1-3 years
Preschooler: 3-6 years
School Ager: 6-12 years
Adolescent: 12-18 years
Adult: 18-64 years
Geriatric: 64 years and older
}
