% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cad_data.R
\name{get_cad_incidents}
\alias{get_cad_incidents}
\title{Import TSTATS from CAD}
\usage{
get_cad_incidents(
  server,
  database,
  UID,
  PWD,
  agency,
  end_date = FALSE,
  start_date = FALSE,
  tzone = "US/Central"
)
}
\arguments{
\item{server}{string}

\item{database}{string}

\item{UID}{string}

\item{PWD}{string}

\item{agency}{string typically "ems01"}

\item{end_date}{end date (defaults to last month end)}

\item{start_date}{start date (defaults to last month start)}

\item{tzone}{Lubridate time zone to target dates/times to. DB stores them in UTC, default target is "US/Central"}
}
\value{

}
\description{
Requires network connection to CAD reporting database.
}
