% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_root.R
\name{get_file_root}
\alias{get_file_root}
\title{Gets the file directory for ESO data.
Generates the file path for ESO data based on the
local system; if in interactive mode gives option
to adjust if needed.}
\usage{
get_file_root(override_interactive = FALSE)
}
\value{
file_root path (as string)
}
\description{
Gets the file directory for ESO data.
Generates the file path for ESO data based on the
local system; if in interactive mode gives option
to adjust if needed.
}
