% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_operational_period_intervals}
\alias{get_operational_period_intervals}
\title{Get list of operational period intervals}
\usage{
get_operational_period_intervals(startDate, endDate, period_length = 12)
}
\arguments{
\item{startDate}{start date POSIXct}

\item{endDate}{end date POSIXct}

\item{period_length}{length in hours of interval}
}
\value{
vector sequence of intervals
}
\description{
Get list of operational period intervals
}
