% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_all_data}
\alias{import_all_data}
\title{import_all_data}
\usage{
import_all_data(table_name)
}
\arguments{
\item{table_name}{Name of ESO table}
}
\value{

}
\description{
Imports all data in file root for a given table name.
}
