% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_cpr}
\alias{import_cpr}
\title{Imports CPR table}
\usage{
import_cpr(year, month = FALSE)
}
\arguments{
\item{month}{}
}
\value{
tibble with CPR data
}
\description{
Imports CPR table
}
