% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_eso_data}
\alias{import_eso_data}
\title{Import ESO Table}
\usage{
import_eso_data(year, month = FALSE, table_name)
}
\arguments{
\item{year}{YYYY string}

\item{month}{MM string OR logi FALSE.}

\item{table_name}{string of table name to find file.}
}
\value{
Tibble with imported data
}
\description{
Works with consistent naming conventions to
simplify importing the monthly data table export.
}
\details{
Requires file_root to be present in the environment.
}
