% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_incidents}
\alias{import_incidents}
\title{Import Incidents with helpful additional fields added}
\usage{
import_incidents(
  year = NULL,
  month = NULL,
  start = NULL,
  end = NULL,
  join_crew = NULL
)
}
\arguments{
\item{year}{Year of the data desired}

\item{month}{Month of the data desired}

\item{start}{Start date}

\item{end}{End date.}

\item{join_crew}{BOOL Adds columns for lead and driver crew names.}
}
\value{

}
\description{
Uses two approaches to import: the year, month option brings
in incidents based on a given year and month. The start and end
option does so based on time.
}
