% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_patients}
\alias{import_patients}
\title{Import Patients table}
\usage{
import_patients(year, month = FALSE)
}
\arguments{
\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
Tibble with imported data
}
\description{
Import Patients table
}
