% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_stroke}
\alias{import_stroke}
\title{Imports Stroke Form Data}
\usage{
import_stroke(year, month = FALSE)
}
\arguments{
\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
Tibble with stroke form data for selected year and month.
}
\description{
Imports Stroke Form Data
}
