% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_treatments}
\alias{import_treatments}
\title{Import complete treatment set}
\usage{
import_treatments(year, month = FALSE)
}
\arguments{
\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
tibble with treatments, dtTreatment
}
\description{
Import complete treatment set
}
