% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_vitals}
\alias{import_vitals}
\title{Import Vitals}
\usage{
import_vitals(year, month = FALSE)
}
\arguments{
\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
Tibble with imported data, vsDate
}
\description{
Import Vitals
}
