% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{is_cardiac_arrest}
\alias{is_cardiac_arrest}
\title{Wrapper for Cardiac Arrest function}
\usage{
is_cardiac_arrest(data, year, month = FALSE)
}
\arguments{
\item{data}{incident table}

\item{year}{Year of data set as "YYYY" string}

\item{month}{Month of data set as "MM" string, or FALSE to look for full year.}
}
\value{
tibble with OHCA field
}
\description{
Wrapper for Cardiac Arrest function
}
