% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{join_crew}
\alias{join_crew}
\title{Add Crew to ESO table}
\usage{
join_crew(.data, year, month = FALSE)
}
\arguments{
\item{.data}{ESO data table (needs to have PatientCareRecordId index)}

\item{month}{}
}
\value{
ESO data table + Lead, Driver column
}
\description{
Adds a column to any ESO data table for Lead and Driver, using Last, First.
}
