% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_utils.R
\name{merge_eso_data}
\alias{merge_eso_data}
\title{Merges two or more files together}
\usage{
merge_eso_data(year, month = FALSE, table_name, ...)
}
\arguments{
\item{year}{YYYY year to look for}

\item{month}{MM month (or FALSE) to look for}

\item{table_name}{Table name to look for}

\item{...}{Arguments to pass onto read_csv for import.}
}
\value{
dataset
}
\description{
Useful when Ad Hoc limits return row length and you end up with A, B, etc. files
}
