% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{rsi}
\alias{rsi}
\title{Did the patient get RSI meds?}
\usage{
rsi(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{table of incidents}

\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
tibble with RSI column added
}
\description{
Did the patient get RSI meds?
}
