% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{rsi_onscene}
\alias{rsi_onscene}
\title{DId the patient get RSI meds before leaving scene?}
\usage{
rsi_onscene(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{tibble of incidents}

\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
tibble iith RSI_onscene
}
\description{
DId the patient get RSI meds before leaving scene?
}
