% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{store_authentication}
\alias{store_authentication}
\title{Stores authentication}
\usage{
store_authentication(cid, username, password)
}
\arguments{
\item{cid}{Company ID}

\item{username}{Username}

\item{password}{Password}
}
\value{

}
\description{
Stores authentication
}
