% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_root.R
\name{windows_file_root}
\alias{windows_file_root}
\title{Get file root on Windows
Gives interactive options for choosing differently.}
\usage{
windows_file_root(user_name, override_interactive = FALSE)
}
\arguments{
\item{user_name}{system user name}

\item{override_interactive}{inherit from above}
}
\value{
file_root path (as string)
}
\description{
Get file root on Windows
Gives interactive options for choosing differently.
}
