% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{blood}
\alias{blood}
\title{Blood}
\usage{
blood(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{tibble of incidents}

\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
tibble with Blood field
}
\description{
Transforms incidents table to include a field "Blood" indicating if the patient received blood products or not.
}
