% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{filtered_incidents}
\alias{filtered_incidents}
\title{Imports a Filtered Incident Set}
\usage{
filtered_incidents(year, month, filter_set, left_join = FALSE)
}
\arguments{
\item{year}{Year of data set as "YYYY" string}

\item{month}{Month of data set as "MM" string, or FALSE to look for full year.}

\item{filter_set}{Tibble of ESO data}

\item{left_join}{optional argument that does a left join to add the filtered set to incidents.}
}
\description{
Imports the Incidents table for a given year and month
but only includes patient care records that match
\code{`PatientCareRecordId`} in the \code{filter_set}.
}
