% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eso_functions.R
\name{import_incidents}
\alias{import_incidents}
\title{Import Incidents}
\usage{
import_incidents(year, month = FALSE)
}
\arguments{
\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
Tibble with imported data, dtDate, month fields
}
\description{
Import Incidents
}
