% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{is_cardiac_arrest}
\alias{is_cardiac_arrest}
\title{Wrapper for Cardiac Arrest function}
\usage{
is_cardiac_arrest(data, year, month = FALSE)
}
\arguments{
\item{data}{incident table}

\item{year}{String with year to select data for.}

\item{month}{optional string with month to select data for. FALSE will load a full year's worth of data.}
}
\value{
tibble with OHCA field
}
\description{
Wrapper for Cardiac Arrest function
}
