% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eso_transformations.R
\name{rsi_onscene}
\alias{rsi_onscene}
\title{DId the patient get RSI meds before leaving scene?}
\usage{
rsi_onscene(the_incidents, year, month = FALSE)
}
\arguments{
\item{the_incidents}{tibble of incidents}

\item{year}{Year of data set as "YYYY" string}

\item{month}{Month of data set as "MM" string, or FALSE to look for full year.}
}
\value{
tibble iith RSI_onscene
}
\description{
DId the patient get RSI meds before leaving scene?
}
