% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uio}
\alias{uio}
\title{User Input/Output}
\usage{
uio(output, wait = FALSE)
}
\arguments{
\item{output}{is the text displayed (output) to}

\item{wait}{specifies whether or not the script}
}
\value{
TRUE (if wait is FALSE) or value entered by the user.
}
\description{
uio is a wrapper function that provides the
ability to interact with the user. If not
in interactive mode, this function stops
execution as a protective measure.
}
